+++ include/GFK/Network/NetworkBuffer.hpp
@@ -18,8 +18,8 @@ public:
 	void WriteUnsignedInt16(unsigned short int i);
 	void WriteSignedInt32(signed int i);
 	void WriteUnsignedInt32(unsigned int i);
-	void WriteSignedInt64(signed long int i);
-	void WriteUnsignedInt64(unsigned long int i);
+	void WriteSignedInt64(signed long long int i);
+	void WriteUnsignedInt64(unsigned long long int i);
 	void WriteFloat32(float f);
 	void WriteFloat64(double d);
 
@@ -29,8 +29,8 @@ public:
 	unsigned short int ReadUnsignedInt16();
 	signed int ReadSignedInt32();
 	unsigned int ReadUnsignedInt32();
-	signed long int ReadSignedInt64();
-	unsigned long int ReadUnsignedInt64();
+	signed long long int ReadSignedInt64();
+	unsigned long long int ReadUnsignedInt64();
 	float ReadFloat32();
 	double ReadFloat64();
 protected:
